feat: Update frontend and backend code

+++ web-services/src/main/java/com/dtaquito_backend/dtaquito_backend/sportspaces/application/internal/commandservices/SportSpacesCommandServiceImpl.java
@@ -2,8 +2,11 @@ package com.dtaquito_backend.dtaquito_backend.sportspaces.application.internal.c
 
 import com.dtaquito_backend.dtaquito_backend.sportspaces.domain.model.aggregates.SportSpaces;
 import com.dtaquito_backend.dtaquito_backend.sportspaces.domain.model.commands.CreateSportSpacesCommand;
+import com.dtaquito_backend.dtaquito_backend.sportspaces.domain.model.events.SportSpacesCreatedEvent;
 import com.dtaquito_backend.dtaquito_backend.sportspaces.domain.services.SportSpacesCommandService;
+import com.dtaquito_backend.dtaquito_backend.sportspaces.infrastructure.persistance.jpa.SportRepository;
 import com.dtaquito_backend.dtaquito_backend.sportspaces.infrastructure.persistance.jpa.SportSpacesRepository;
+import com.dtaquito_backend.dtaquito_backend.users.domain.exceptions.InvalidRoleTypeException;
 import com.dtaquito_backend.dtaquito_backend.users.infrastructure.persistance.jpa.UserRepository;
 import org.springframework.stereotype.Service;
 
@@ -14,17 +17,24 @@ public class SportSpacesCommandServiceImpl implements SportSpacesCommandService 
 
     private final SportSpacesRepository sportSpacesRepository;
     private final UserRepository userRepository;
+    private final SportRepository sportRepository;
 
-    public SportSpacesCommandServiceImpl(SportSpacesRepository sportSpacesRepository, UserRepository userRepository) {
+    public SportSpacesCommandServiceImpl(SportSpacesRepository sportSpacesRepository, UserRepository userRepository
+    , SportRepository sportRepository) {
         this.sportSpacesRepository = sportSpacesRepository;
         this.userRepository = userRepository;
+        this.sportRepository = sportRepository;
     }
 
     @Override
     public Optional<SportSpaces> handle(Long userId, CreateSportSpacesCommand command) {
         var user = userRepository.findById(userId)
                 .orElseThrow(() -> new IllegalArgumentException("User not found"));
-        var sportSpaces = new SportSpaces(command, user);
+        var sport = sportRepository.findById((command.sportId()));
+        if(!sportRepository.existsById(command.sportId())){
+            throw new InvalidRoleTypeException("Sport id: " + command.sportId() + " already exists");
+        }
+        var sportSpaces = new SportSpaces(command, user, sport.get());
         var createdSportSpaces = sportSpacesRepository.save(sportSpaces);
         return Optional.of(createdSportSpaces);
     }
@@ -40,6 +50,14 @@ public class SportSpacesCommandServiceImpl implements SportSpacesCommandService 
 
     @Override
     public void handleDelete(Long id) {
+        if (!sportSpacesRepository.existsById(id)) {
+            throw new IllegalArgumentException("SportSpaces not found");
+        }
         sportSpacesRepository.deleteById(id);
+    }
+
+    @Override
+    public void handleSportSpacesCreatedEvent(SportSpacesCreatedEvent event) {
+        System.out.println("SportSpacesCreatedEvent received for sportSpaces ID: " + event.getId());
     }
 }
